+++ Education.Application/Services/CourseServices/CourseService.cs
@@ -110,7 +110,7 @@ namespace Education.Application.Services.CourseServices
 				return new ApiResponse<string>(500, "An unexpected error occurred while processing the Delete Course");
 			}
 		}
-		public async Task<ApiResponse<CourseRespondDto>> UpdateCourse(UpdateCourseDto coursesDto)
+		public async Task<ApiResponse<CourseRespondDto>> UpdateCourse(int courseId,UpdateCourseDto coursesDto)
 		{
 			bool IsNewPhotoUpload = false;
 			if (coursesDto is null)
@@ -122,18 +122,30 @@ namespace Education.Application.Services.CourseServices
 				//if (category is null)
 					//new ApiResponse<CourseRespondDto>(404, $"Category Id :{coursesDto.CategoriesId}  Is Not Found");
 
-				var course = await courseRepository.GetEntityAsync(e => e.CoursesId == coursesDto.CourseId && !e.IsDeleted, Includes, true);
+				var course = await courseRepository.GetEntityAsync(e => e.CoursesId == courseId && !e.IsDeleted, Includes, true);
 				if (course is null)
-					return new ApiResponse<CourseRespondDto>(404, $"Course With Id :{coursesDto.CourseId}  Is Not Found");
+					return new ApiResponse<CourseRespondDto>(404, $"Course With Id :{courseId}  Is Not Found");
 
-				imageService.DeleteImage($"{CourseFolderName}/{course.CourseImage}");
+				
 
-				var imageResult = await imageService.UploadImage(coursesDto.CourseImage, CourseFolderName);
-				if (!imageResult.IsUploaded)
-					return new ApiResponse<CourseRespondDto>(400, imageResult.ErrorMessage);
+				Console.WriteLine(coursesDto.CourseImage is null);
 
-				course.CourseImage = imageResult.ImageName;
-				IsNewPhotoUpload = true;
+				if (coursesDto.CourseImage is not null)
+				{
+					var imageResult = await imageService.UploadImage(coursesDto.CourseImage, CourseFolderName);
+					if (!imageResult.IsUploaded)
+						return new ApiResponse<CourseRespondDto>(400, imageResult.ErrorMessage);
+
+                    imageService.DeleteImage($"{CourseFolderName}/{course.CourseImage}");
+
+                    course.CourseImage = imageResult.ImageName;
+					IsNewPhotoUpload = true;
+				}
+				
+
+
+
+
 				if (!Enum.TryParse<CourseStatus>(coursesDto.CourseStatus, true, out var courseStatus))
 					return new ApiResponse<CourseRespondDto>(400, "Invalid Course Status Value");
 
@@ -141,7 +153,7 @@ namespace Education.Application.Services.CourseServices
 				course.LastUpdateOn = DateTime.Now;
 				MapDTOToCourse(course, coursesDto, null);
 
-				courseRepository.Update(course);
+				//courseRepository.Update(course);
 				await courseRepository.SaveChangesAsync();
 
 				var CourseRsepond = MapCourseToDTO(course);
@@ -151,7 +163,7 @@ namespace Education.Application.Services.CourseServices
 			{
 				if (IsNewPhotoUpload)
 				{
-					var course = await courseRepository.GetByIdAsync(coursesDto.CourseId);
+					var course = await courseRepository.GetByIdAsync(courseId);
 					imageService.DeleteImage($"{CourseFolderName}/{course.CourseImage}");
 				}
 				return new ApiResponse<CourseRespondDto>(500, "An unexpected error occurred while processing the Course .");
@@ -189,19 +201,20 @@ namespace Education.Application.Services.CourseServices
 			course.Price = dto.Price;
 			course.DiscountPercentage = dto.DiscountPercentage;
 			course.IsSequentialWatch = dto.IsSequentialWatch;
+			//course.CourseStatus = dto.CourseStatus;
 		}
 
-		public async Task<ApiResponse<string>> ChangeCourseAccess(ChangeAccessDto changeAccessDto)
+		public async Task<ApiResponse<string>> ChangeCourseAccess(int courseId,ChangeAccessDto changeAccessDto)
 		{
 			if (changeAccessDto is null)
 				return new ApiResponse<string>(400, "Course data is missing");
 			try
 			{
-				var course = await courseRepository.GetByIdAsync(changeAccessDto.CourseId);
+				var course = await courseRepository.GetByIdAsync(courseId);
 				if (course is null)
-					return new ApiResponse<string>(404, $"Course With Id :{changeAccessDto.CourseId}  Is Not Found");
+					return new ApiResponse<string>(404, $"Course With Id :{courseId}  Is Not Found");
 				if (course.IsDeleted)
-					return new ApiResponse<string>(404, $"Course With Id :{changeAccessDto.CourseId}  Is Deleted");
+					return new ApiResponse<string>(404, $"Course With Id :{courseId}  Is Deleted");
 
 				string Oldstatus = course.IsFree ? "Free" : "Paid";
